김초원 : [DP] 가장 긴 증가하는 부분 수열

+++ 백준 알고리즘/[DP] 이친수/[DP] 이친수.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 using namespace std;
 
-int d[91][2];
+long long d[91][2];
 
 int main()
 {
